+++ junitTest/src/junitTest/CarTests.java
@@ -22,12 +22,12 @@ class CarTests {
 
 	@Test
 	void CanComputeEconomyImperial() {
-		assertEquals(31.2, Car.ComputeEconomy((short)312, (short)10));
+		assertEquals((float)31.2, Car.ComputeEconomy((short)312, (short)10));
 	}
 	
 	@Test
 	void CanComputeEconomyMetric() {
-		assertEquals(13.210526, Car.ComputeEconomy(502, 38));
+		assertEquals((float)13.210526, Car.ComputeEconomy(502, 38));
 	}
 
 }
